Consolidate DB connections / pool
