+++ src/main/kotlin/au/gov/api/DB.kt
@@ -1,0 +1,33 @@
+package au.gov.api.db
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Bean
+import org.springframework.stereotype.Component
+import java.sql.Connection
+import java.sql.SQLException
+import javax.sql.DataSource
+import com.zaxxer.hikari.HikariConfig
+import com.zaxxer.hikari.HikariDataSource
+
+@Component
+class DB{
+
+    @Value("\${spring.datasource.url}")
+    var dbUrl: String? = "jdbc:postgresql://localhost:5432/postgres" 
+
+    @Bean
+    @Throws(SQLException::class)
+    fun dataSource(): DataSource? {
+        if (dbUrl?.isEmpty() ?: true) {
+            return HikariDataSource()
+        } else {
+            val config = HikariConfig()
+            config.jdbcUrl = dbUrl
+            try {
+                return HikariDataSource(config)
+            } catch (e: Exception) {
+                return null
+            }
+        }
+    }
+}
