+++ src/main/kotlin/au/gov/api/registration/RegistrationManager.kt
@@ -21,8 +21,6 @@ import javax.sql.DataSource
 @Component
 class RegistrationManager{
 
-        @Value("\${spring.datasource.url}")
-        private var dbUrl: String? = null
 
         @Autowired
         private lateinit var dataSource: DataSource
@@ -150,19 +148,4 @@ class RegistrationManager{
         }
 
 
-        @Bean
-        @Throws(SQLException::class)
-        fun dataSource(): DataSource? {
-            if (dbUrl?.isEmpty() ?: true) {
-                return HikariDataSource()
-            } else {
-                val config = HikariConfig()
-                config.jdbcUrl = dbUrl
-                try {
-                    return HikariDataSource(config)
-                } catch (e: Exception) {
-                    return null
-                }
-            }
-        }
 }
